Split more tests to speed up Travis run
