+++ scripts/runTestsOnTravis.sh
@@ -32,38 +32,54 @@ elif [ "$1" = "js_tests" ]; then
 
     # Run JetStream non-clustered tests. By convention, all JS tests start
     # with `TestJetStream`. We exclude the clustered and super-clustered
-    # tests by using the `skip_js_cluster_tests` and `skip_js_super_cluster_tests`
-    # build tags.
+    # tests by using the `skip_js_cluster_tests`, `skip_js_cluster_tests_2`
+    # and `skip_js_super_cluster_tests` build tags.
 
-    go test -race -v -run=TestJetStream ./server -tags=skip_js_cluster_tests,skip_js_super_cluster_tests -count=1 -vet=off -timeout=30m -failfast
+    go test -race -v -run=TestJetStream ./server -tags=skip_js_cluster_tests,skip_js_cluster_tests_2,skip_js_super_cluster_tests -count=1 -vet=off -timeout=30m -failfast
 
-elif [ "$1" = "js_cluster_tests" ]; then
+elif [ "$1" = "js_cluster_tests_1" ]; then
 
     # Run JetStream clustered tests. By convention, all JS cluster tests
-    # start with `TestJetStreamCluster`.
+    # start with `TestJetStreamCluster`. Will run the first batch of tests,
+    # excluding the ones with the tag skip_js_cluster_tests_2.
 
-    go test -race -v -run=TestJetStreamCluster ./server -count=1 -vet=off -timeout=30m -failfast
+    go test -race -v -run=TestJetStreamCluster ./server -tags=skip_js_cluster_tests_2 -count=1 -vet=off -timeout=30m -failfast
+
+elif [ "$1" = "js_cluster_tests_2" ]; then
+
+    # Run JetStream clustered tests. By convention, all JS cluster tests
+    # start with `TestJetStreamCluster`. Will run the second batch of tests,
+    # excluding the ones with the tag skip_js_cluster_tests.
+
+    go test -race -v -run=TestJetStreamCluster ./server -tags=skip_js_cluster_tests -count=1 -vet=off -timeout=30m -failfast
 
 elif [ "$1" = "js_super_cluster_tests" ]; then
 
     # Run JetStream super clustered tests. By convention, all JS super cluster
-    # tests with `TestJetStreamSuperCluster`.
+    # tests start with `TestJetStreamSuperCluster`.
 
     go test -race -v -run=TestJetStreamSuperCluster ./server -count=1 -vet=off -timeout=30m -failfast
 
 elif [ "$1" = "js_chaos_tests" ]; then
 
-    # Run JetStream chaos tests. By convention, all JS cluster chaos
-    # tests with `TestJetStreamChaos`.
+    # Run JetStream chaos tests. By convention, all JS cluster chaos tests
+    # start with `TestJetStreamChaos`.
 
     go test -race -v -p=1 -run=TestJetStreamChaos ./server -tags=js_chaos_tests -count=1 -vet=off -timeout=30m -failfast
+
+elif [ "$1" = "mqtt_tests" ]; then
+
+    # Run MQTT tests. By convention, all MQTT tests start with `TestMQTT`.
+
+    go test -race -v -run=TestMQTT ./server -count=1 -vet=off -timeout=30m -failfast
 
 elif [ "$1" = "srv_pkg_non_js_tests" ]; then
 
     # Run all non JetStream tests in the server package. We exclude the
-    # JS tests by using the `skip_js_tests` build tag.
+    # JS tests by using the `skip_js_tests` build tag and MQTT tests by
+    # using the `skip_mqtt_tests`
 
-    go test -race -v -p=1 ./server/... -tags=skip_js_tests -count=1 -vet=off -timeout=30m -failfast
+    go test -race -v -p=1 ./server/... -tags=skip_js_tests,skip_mqtt_tests -count=1 -vet=off -timeout=30m -failfast
 
 elif [ "$1" = "non_srv_pkg_tests" ]; then
