+++ server/mqtt_test.go
@@ -11,6 +11,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//go:build !skip_mqtt_tests
+// +build !skip_mqtt_tests
+
 package server
 
 import (
